correct imports and test settings
