+++ tests/aswwu/behaviors/auth/auth_tests.py
@@ -1,0 +1,6 @@
+from tests.aswwu.behaviors.auth.subtests import send_post_verify
+from tests.conftest import testing_server
+
+
+def test_post_verify(testing_server):
+    send_post_verify()
